refactor: move engine into game, keep both classes
