+++ cpp/input.cpp
@@ -4,7 +4,7 @@
 
 #include <stdint.h>
 
-#include "engine.h"
+#include "game.h"
 #include "callbacks.h"
 #include "util/timer.h"
