+++ cpp/terrain/terrain.h
@@ -13,7 +13,6 @@
 #include "terrain_chunk.h"
 #include "terrain_object.h"
 #include "../assetmanager.h"
-#include "../engine.h"
 #include "../texture.h"
 #include "../coord/camgame.h"
 #include "../coord/chunk.h"
@@ -25,6 +24,7 @@
 namespace openage {
 
 class TerrainChunk;
+class Engine;
 
 /**
  * type that for terrain ids.
